factor out effect composer factory
